% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFormat.R
\name{dataFormat}
\alias{dataFormat}
\title{Function to format the .csv data generated from camtrapR}
\usage{
dataFormat(
  data_Spp,
  data_CT,
  speciesCol,
  stationCol,
  setupCol,
  retrievalCol,
  CTDateFormat
)
}
\arguments{
\item{data_Spp}{Matrix of species}

\item{data_CT}{Matrix of camera-traps}

\item{speciesCol}{Spp}

\item{stationCol}{CT}

\item{setupCol}{Initial}

\item{retrievalCol}{End}

\item{CTDateFormat}{Date}
}
\description{
Function to format the .csv data generated from camtrapR
}
\author{
"SMandujanoR"
}
