% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi_Grid.R
\name{mi_Grid}
\alias{mi_Grid}
\title{Function to create a camera-trap grid}
\usage{
mi_Grid(n1, n2, CTx, CTy, CTdist)
}
\arguments{
\item{n1}{Number of camera-traps in columns}

\item{n2}{Number of camera-traps in rows}

\item{CTx}{Define coordinates X in UTMs}

\item{CTy}{Define coordinates Y in UTMs}

\item{CTdist}{Define distance among camera-traps}
}
\description{
Function to create a camera-trap grid
}
\author{
SMandujanoR
}
