% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rai.R
\name{RAI}
\alias{RAI}
\alias{rai}
\title{Main function for RAI regression.}
\usage{
rai(theData, theResponse, alpha = 0.05, alg = "rai", r = 0.8,
  poly = (alg != "RH"), searchType = "breadth", m = 500,
  sigma = "step", rmse = NA, df = NA, omega = alpha, reuse = (alg
  == "RH"), maxTest = Inf, verbose = F, save = T)
}
\arguments{
\item{theData}{matrix of covariates.}

\item{theResponse}{response vector stored as a single column matrix.}

\item{alpha}{level of procedure.}

\item{alg}{algorithm can be one of "rai", "raiPlus", or "RH" (Revisiting
Holm).}

\item{r}{RAI rejects tests which increase R^2 by a factor r^s, where s is the
epoch.}

\item{poly}{logical. Should the algorithm look for higher-order polynomials?}

\item{searchType}{A character string specifying the prioritization of
higher-order polynomials. One of "breadth" (more base features) or "depth"
(higher orders).}

\item{m}{number of observations used in subsampling for variance inflation
factor estimate of r.squared.}

\item{sigma}{type of error estimate used; one of "ind" or "step". If "ind",
you must provide a numeric value for rmse and df.}

\item{rmse}{user provided value for rmse. Must be used with sigma="ind".}

\item{df}{degrees of freedom for user specified rmse. Must be used with
sigma="ind".}

\item{omega}{return from rejecting a test in Alpha-Investing (<= alpha).}

\item{reuse}{logical. Should repeated tests of the same covariate be
considered a test of the same hypothsis? reusing wealth isn't implemented
for RAI or RAIplus as the effect is negligible.}

\item{maxTest}{maximum number of tests.}

\item{verbose}{logical. Should auction output be prited?}

\item{save}{logical. Should the auction results be saved? If TRUE, returns a
summary matrix.}
}
\value{
A list which includes the following components: \item{formula}{final
  model formula.} \item{y}{response.} \item{X}{model matrix from final
  model.} \item{experts}{list of experts.} \item{theModelFeatures}{list of
  feature names in the final model.} \item{options}{options given to RAI:
  algorithm, searchType, r, poly}
}
\description{
RAI is a wrapper function to create the required objects and
  manage the input/ouptut of the runAuction function. Using poly=F is an
  efficient and statistically valid way to run stepwise regression.
}
\examples{
  data("CO2")
  theResponse = CO2$uptake
  theData = CO2[ ,-5]
  rai_out = rai(theData, theResponse)
  summarise_rai(rai_out)
  raiPlus_out = rai(theData, theResponse, alg="raiPlus")
  summarise_rai(raiPlus_out)$cost_raiPlus
}
