% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIalt.R
\name{RAIalt}
\alias{RAIalt}
\title{Function to calculate the alternative RAI model}
\usage{
RAIalt(new.mat, max)
}
\arguments{
\item{new.mat}{Matrix with data of camera-traps}

\item{max}{Select maxime value to plot Y-axes}
}
\description{
Function to calculate the alternative RAI model
}
\author{
"SMandujanoR"
}
