% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIfinal.R
\name{RAIfinal}
\alias{RAIfinal}
\title{Function to create final Table}
\usage{
RAIfinal(new.mat)
}
\arguments{
\item{new.mat}{Matrix with data of camera-traps}
}
\description{
Function to create final Table
}
\author{
SMandujanoR
}
