% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIglm.R
\name{RAIglm}
\alias{RAIglm}
\title{Function GLM Poisson model with "offsets"}
\usage{
RAIglm(new.mat, family = family)
}
\arguments{
\item{new.mat}{Matrix with data of camera-traps}

\item{family}{Choise among "Gaussian", "Poisson", "quasipoisson" distribution}
}
\description{
Function GLM Poisson model with "offsets"
}
\author{
SMandujanoR
}
