% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIglmSp.R
\name{RAIglmSp}
\alias{RAIglmSp}
\title{Function to GLM-Poisson for selected species and analysis of one covariate}
\usage{
RAIglmSp(Species, covs, family)
}
\arguments{
\item{Species}{Select the species}

\item{covs}{Define covariates}

\item{family}{Define the error distribution}
}
\description{
Function to GLM-Poisson for selected species and analysis of one covariate
}
\author{
SMandujanoR
}
