% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIgral.R
\name{RAIgral}
\alias{RAIgral}
\title{Function to calculate the classic or general RAI model}
\usage{
RAIgral(new.mat, Species, parRow, parCol, pointSize)
}
\arguments{
\item{new.mat}{Matrix with data of camera-traps}

\item{Species}{Select the species to plot}

\item{parRow}{Number of rows}

\item{parCol}{Number of columns}

\item{pointSize}{Select the size of the circles}

\item{RAIgral}{General RAI}
}
\description{
Function to calculate the classic or general RAI model
}
\author{
"SMandujanoR"
}
