% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAIkernelSP.R
\name{RAIkernelSp}
\alias{RAIkernelSp}
\title{Function to kernel utilization distribution for selected species}
\usage{
RAIkernelSp(species, S, HR, pointSize, proje, mycolors)
}
\arguments{
\item{species}{x}

\item{S}{s}

\item{HR}{s}

\item{pointSize}{s}

\item{proje}{s}

\item{mycolors}{x}
}
\description{
Function to kernel utilization distribution for selected species
}
\author{
SMandujanoR
}
