% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAImapSp.R
\name{RAImapSp}
\alias{RAImapSp}
\title{Function to graph the RAI of selected species}
\usage{
RAImapSp(map, species, pointSize, mycolors)
}
\arguments{
\item{map}{Introduce map}

\item{species}{Select the species}

\item{pointSize}{X}

\item{mycolors}{Select}
}
\description{
Function to graph the RAI of selected species
}
\author{
SMandujanoR
}
